Renamed main component (PlanningOverview) to index.tsx
